refactor(deploy-link): the deploy link moves to a package both ends can read
The link the deploy screen validates its search against was the dashboard's
alone, so nothing else could write one without spelling its parameters out a
second time. It, and the binary url checks it reads with, are now a package.
