+++ apps/dashboard/package.json
@@ -15,6 +15,7 @@
     "@repo/api-client": "workspace:*",
     "@repo/app-operations": "workspace:*",
     "@repo/binary-handoff": "workspace:*",
+    "@repo/deploy-link": "workspace:*",
     "@repo/protocol": "workspace:*",
     "@repo/ui": "workspace:*",
     "@tanstack/react-devtools": "^0.10.9",
