+++ apps/dashboard/src/components/deploy/deploy-form.tsx
@@ -1,3 +1,4 @@
+import type { DeploySuggestion } from '@repo/deploy-link';
 import { Button } from '@repo/ui/components/button';
 import { useStore } from '@tanstack/react-form';
 import { AdvancedConfiguration } from '#components/deploy/advanced-configuration.tsx';
@@ -5,7 +6,6 @@ import { DeployBinaryField } from '#components/deploy/deploy-binary-field.tsx';
 import { asksForVariables, DeployConfiguration } from '#components/deploy/deploy-configuration.tsx';
 import { MinimalBinaryField } from '#components/deploy/minimal-binary-field.tsx';
 import { binaryName } from '#lib/binary-source.ts';
-import type { DeploySuggestion } from '#lib/deploy-link.ts';
 import { type DeployFormValues, useDeployForm } from '#lib/hooks/use-deploy-form.ts';
 import type { AppSummary } from '#queries/apps.ts';
