+++ apps/dashboard/src/lib/hooks/use-deploy-link.ts
@@ -1,4 +1,4 @@
-import type { DeployLink } from '#lib/deploy-link.ts';
+import type { DeployLink } from '@repo/deploy-link';
 import { Route as DeployRoute } from '#routes/deploy.tsx';
 
 export function useDeployLink(): DeployLink {
