+++ apps/dashboard/tests/lib/binary-source.test.ts
@@ -1,12 +1,5 @@
 import { expect, test } from 'bun:test';
-import {
-  binaryName,
-  fetchedUrl,
-  namedByUrl,
-  pickedFile,
-  refusedUrl,
-  sourceFromUrl,
-} from '#lib/binary-source.ts';
+import { binaryName, fetchedUrl, pickedFile, sourceFromUrl } from '#lib/binary-source.ts';
 
 const URL_SOURCE = 'https://github.com/me/app/releases/download/v1/my-server';
 const FILE = new File([], 'my-server');
@@ -30,15 +23,4 @@ test('an empty box is not a binary', () => {
   expect(sourceFromUrl('')).toBeUndefined();
   expect(sourceFromUrl('   ')).toBeUndefined();
   expect(sourceFromUrl(` ${URL_SOURCE} `)).toEqual({ url: URL_SOURCE });
-});
-
-/**
- * Said here rather than by the api, because the api is the end that follows the url: a mistake it
- * would answer with is one the owner would otherwise wait through a deploy to hear about.
- */
-test('a url nibrun could not fetch a binary from is refused before the deploy', () => {
-  expect(refusedUrl(URL_SOURCE)).toBeUndefined();
-  expect(refusedUrl('http://releases.test/my-server')).toContain('https');
-  expect(refusedUrl('https://releases.test/downloads/')).toContain('name');
-  expect(namedByUrl('https://releases.test/../etc/passwd')).toBe('passwd');
 });
