+++ packages/deploy-link/src/index.ts
@@ -1,0 +1,9 @@
+/** biome-ignore-all lint/performance/noBarrelFile: index is the only allowed file where we can export other files */
+
+export { namedByUrl, refusedUrl } from '#binary-url.ts';
+export {
+  type DeployLink,
+  type DeploySuggestion,
+  deployLink,
+  deploySuggestion,
+} from '#link.ts';
